requirements.txt
# web
Flask==2.2.2

# environment
python-dotenv==0.21.0
